dark theme, photo taking, values as integers, app_spacing, app_sizes. validations on numbers.

+++ lib/src/features/foods/data/foods_repository.dart
@@ -23,9 +23,9 @@ class FoodsRepository {
     String? id,
     required String name,
     String? photoPath,
-    double carbsGrams = 0,
-    double sodiumMg = 0,
-    double caffeineMg = 0,
+    int carbsGrams = 0,
+    int sodiumMg = 0,
+    int caffeineMg = 0,
     String? notes,
   }) {
     return _dao.upsert(
